feat: Welcome 페이지 엔드포인트 변경
